Use Long for the moves array
I'm not sure what the third element in the moves array is, but it can be
greater than an integer.

com.squareup.moshi.JsonDataException: Expected an int but was 4203354345 at path $.active_games[0].json.moves[63][2]

Fixes #8 - Games don't show up

+++ app/src/main/java/com/ogs/Models.kt
@@ -177,7 +177,7 @@ class Gamedata {
     var phase: String? = null
     //var history: List<Any>? = null
     var initial_player: String? = null
-    var moves: List<List<Int>>? = null
+    var moves: List<List<Long>>? = null
     var allow_self_capture: Boolean? = null
     var automatic_stone_removal: Boolean? = null
     var free_handicap_placement: Boolean? = null
